refactor message responses & instantiate schema/model definitions
