+++ initializers/index.ts
@@ -1,1 +1,2 @@
 import './mongo';
+import './schema';
