+++ initializers/schema.ts
@@ -1,0 +1,9 @@
+import mongoose from 'mongoose';
+import '../models';
+import logger from '../utils/Logger';
+
+export function main(): void {
+    logger.info('Initializing schema definitions...');
+}
+
+main();
