+++ models/index.ts
@@ -1,0 +1,5 @@
+/**
+ * Import any models that should be instantiated here.
+ * Their schema's will be built the first time
+ */
+import './User.model';
